% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{anomalize_ss_anom_at}
\alias{anomalize_ss_anom_at}
\title{*_ss_anom_la Anomaly Detection}
\usage{
anomalize_ss_anom_at(fot_input_tbl, grp_vars, time_var, var_col)
}
\arguments{
\item{fot_input_tbl}{table output by \code{\link[=compute_fot]{compute_fot()}} for *_ss_anom_at analyses}

\item{grp_vars}{the variables that should be preserved in the cross join}

\item{time_var}{the column with time information}

\item{var_col}{the column with the numerical statistic of interest for the euclidean
distance computation}
}
\value{
one dataframe with all columns from the original input table
plus the columns needed for timetk output generated by the
\code{anomalize} function
}
\description{
For analyses where the \code{time_period} is smaller than a year, this function
will execute \code{\link[timetk:anomalize]{timetk::anomalize()}} to identify outliers in the time series
using STL regression. For year-level analyses, the same input table will be
returned and a different anomaly detection method will be used at the *_output
stage
}
