% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_site.R
\name{check_site_type}
\alias{check_site_type}
\title{Check Site Type}
\usage{
check_site_type(cohort, multi_or_single_site)
}
\arguments{
\item{cohort}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}}

\item{multi_or_single_site}{\emph{string} || defaults to \code{single}

A string, either \code{single} or \code{multi}, indicating whether a single-site or
multi-site analysis should be executed}
}
\value{
If \code{multi_or_single_site = single} but multiple sites are provided, the cohort table
is returned with a summary site column set to \code{combined} so all sites will be treated
as one group. Otherwise, the existing site column is returned as-is.
If an illogical parameter combination is supplied, the function will
return an error with recommendations on how to remedy the issue.
}
\description{
This function will evaluate the number of distinct values in the \code{site} column in
the provided cohort table and determine how that compares to the provided
\code{multi_or_single_site} designation.
}
\examples{
## Create sample cohort
cohort_sample <- dplyr::tibble(site = c('Site A', 'Site B', 'Site C'),
                               person_id = c(1,2,3))

## If number of sites & indicated multi/single site match, output same table
check_site_type(cohort = cohort_sample,
                multi_or_single_site = 'multi')

## If multiple sites but single site indicated, create site_summ column
check_site_type(cohort = cohort_sample,
                multi_or_single_site = 'single')

}
