% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_prep.R
\name{cohort_codeset_label_omop}
\alias{cohort_codeset_label_omop}
\title{Custom Codeset-Based Flags -- OMOP}
\usage{
cohort_codeset_label_omop(cohort_tbl, codeset_meta)
}
\arguments{
\item{cohort_tbl}{table of cohort members with at least \code{person_id}, \code{start_date}, and \code{end_date}}

\item{codeset_meta}{a CSV file with metadata relating to a codeset with customized group labels

\if{html}{\out{<div class="sourceCode">}}\preformatted{                this file should have `table`, `column`, and `file_name` columns
}\if{html}{\out{</div>}}}
}
\description{
Custom Codeset-Based Flags -- OMOP
}
