% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_prep.R
\name{compute_age_groups}
\alias{compute_age_groups}
\title{compute age at cohort entry}
\usage{
compute_age_groups(cohort_tbl, person_tbl, age_groups)
}
\arguments{
\item{cohort_tbl}{table of cohort members with at least \code{person_id}, \code{start_date}, and \code{end_date}}

\item{person_tbl}{the CDM person table}

\item{age_groups}{a csv file (template found in specs folder) where the user defines the minimum and maximum
age allowed for a group and provides a string name for the group}
}
\value{
\code{cohort_tbl} with the age at cohort entry and age group for each patient
}
\description{
compute age at cohort entry
}
