% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{compute_at_cross_join}
\alias{compute_at_cross_join}
\title{Create a cross-joined master table for variable reference}
\usage{
compute_at_cross_join(
  cj_tbl,
  cj_var_names = c("site", "concept_id"),
  join_type = "left"
)
}
\arguments{
\item{cj_tbl}{a table with the results of a longitudinal analysis}

\item{cj_var_names}{a vector with the names of variables that should be used as the "anchor"
of the cross join where all combinations of the variables should be
present in the final table}

\item{join_type}{the type of join that should be performed at the end of the function
left is used for multi-site anomaly (euclidean distance) while full
is used for single site anomaly (timetk package)}
}
\value{
one data frame with all combinations of the variables from cj_var_names with their
associated facts from the original cj_tbl input
}
\description{
Create a cross-joined master table for variable reference
}
\keyword{internal}
