% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{compute_dist_anomalies}
\alias{compute_dist_anomalies}
\title{Hotspots Anomaly Detection Eligibility Determination}
\usage{
compute_dist_anomalies(df_tbl, grp_vars, var_col, denom_cols)
}
\arguments{
\item{df_tbl}{\emph{tabular input} || \strong{required}

A dataframe with at least one numerical variable & any relevant variables
needed for grouping}

\item{grp_vars}{\emph{string or vector} || \strong{required}

The variable(s) to be used as grouping variables in the analysis}

\item{var_col}{\emph{string} || \strong{required}

The variable with the numerical statistic of interest for the euclidean
distance computation}

\item{denom_cols}{\emph{string or vector} || \strong{required}

The variable containing a denominator or any other variables that
should be preserved without nulls after a cross_join takes place}
}
\value{
This function will return the original \code{df_tbl} with the addition of the
summary statistics used in the eligibility computation and a flag indicating
whether a given variable (based on the grp_vars) is eligible for anomaly
detection analysis. This table can then be passed into \code{detect_outliers}
to identify anomalous values.
}
\description{
This function will, for each group in a dataframe, identify groups that are eligible
for anomaly detection analysis by examining the values in the \code{var_col}.
The following conditions will disqualify a group from the anomaly detection analysis:
(1) Mean < 0.02 or Median < 0.01
(2) Mean value < 0.05 and range < 0.01
(3) Coefficient of variance < 0.1 and sample size < 11
If no groups meet this criteria, a warning will display in the console indicating
that no groups were eligible.
}
\examples{
# create sample input (modeled after EVP)
sample_ms_input <- dplyr::tibble('site' = c('Site A', 'Site A', 'Site A',
                                            'Site A', 'Site B', 'Site B',
                                            'Site B', 'Site B'),
                                 'variable' = c('dx', 'dx', 'drug', 'drug',
                                                'dx', 'dx', 'drug', 'drug'),
                                 'count' = c(100, 140, 39, 42, 137, 111,
                                             12, 15),
                                 'total_var' = c(1000, 1000, 200, 200, 1500,
                                                 1500, 100, 100))
# execute the full analysis, including compute_dist_anomalies and
# detect_outliers
anomaly_output1 <- compute_dist_anomalies(df_tbl = sample_ms_input,
                                          grp_vars = 'variable',
                                          var_col = 'count',
                                          denom_cols = 'total_var')

anomaly_output1

anomaly_output2 <- detect_outliers(df_tbl = anomaly_output1,
                                   column_analysis = 'count',
                                   column_variable = 'variable')

anomaly_output2

}
