% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{compute_dist_mean_median}
\alias{compute_dist_mean_median}
\title{Compute Distance from Mean & Median}
\usage{
compute_dist_mean_median(tbl, grp_vars, var_col, num_sd, num_mad)
}
\arguments{
\item{tbl}{table with at least the vars specified in \code{grp_vars} and \code{var_col}}

\item{grp_vars}{variables to group by when computing summary statistics}

\item{var_col}{column to compute summary statistics on}

\item{num_sd}{(integer) number of standard deviations away from the mean
from which to compute the sd_lower and sd_upper columns}

\item{num_mad}{(integer) number of median absolute deviations away from the median
from which to compute the mad_lower and mad_upper columns}
}
\value{
a table where, for each group in \code{grp_vars}, the following statistics are computed
based on the \code{var_col}: mean, sd, sd_lower, sd_upper, mad, mad_lower, mad_upper,
anomaly_yn (indicator of whether data point is +/- num_sd from mean or data point is > 90th percentile),
abs_diff_mean (absolute value of difference between mean for group and observation),
abs_diff_median (absolute value of difference between median for group and observation),
n_mad (the number of MAD away from the median where the data point falls)
}
\description{
This function will, for the values in a given \code{var_col}, compute the
distance between that value and the mean & median values for each group
established by the \code{grp_vars} parameter.
}
