% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{compute_euclidean}
\alias{compute_euclidean}
\title{Compute Euclidean Distance}
\usage{
compute_euclidean(ms_tbl, output_var, grp_vars = c("site", "concept_id"))
}
\arguments{
\item{ms_tbl}{output from \code{\link[=compute_dist_mean_median]{compute_dist_mean_median()}} where the cross-joined table from
\code{\link[=compute_at_cross_join]{compute_at_cross_join()}} is used as input}

\item{output_var}{the output variable that should be used to compute the Euclidean distance
i.e. a count or proportion}

\item{grp_vars}{vector of grouping variables; the euclidean distance will be computed per group}
}
\value{
one dataframe with all variables from ms_tbl with the addition of columns with a site Loess
value and a site Euclidean distance value
}
\description{
Compute Euclidean Distance
}
\keyword{internal}
