% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_loops.R
\name{compute_fot}
\alias{compute_fot}
\title{Compute Facts Over Time}
\usage{
compute_fot(
  cohort,
  check_func,
  site_col,
  reduce_id = NULL,
  time_period = "year",
  time_span = c("2012-01-01", "2020-12-31"),
  site_list
)
}
\arguments{
\item{cohort}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}

Note that the start and end dates included in this table will be used to
limit the search window for the analyses in this module.}

\item{check_func}{\emph{function} || \strong{required}

The function that should be executed within each time period in the loop. This
parameter should be structured as the following, where \code{dat} is the input data
for the function:

\code{function(dat){check_function(param1 = dat, param2 = param2_input, ..., paramX = paramX_input)}}

Make sure to include all parameters required for the original function where
default values are not being used.}

\item{site_col}{\emph{string} || \strong{required}

The name of the column in the cohort table that contains the site names. This
will typically be either \code{site} or \code{site_summ}}

\item{reduce_id}{\emph{string} || defaults to \code{NULL}

If the function provided in \code{check_func} returns a list of tables, this parameter
should be the name of the column that should be used to reduce the tables into
one dataframe (via \link[dplyr:bind_rows]{dplyr::bind_rows})}

\item{time_period}{\emph{string} || defaults to \code{year}

A string indicating the distance between dates within the specified time_span.
Defaults to \code{year}, but other time periods such as \code{month} or \code{week} are
also acceptable}

\item{time_span}{\emph{vector - length 2} || defaults to \code{c('2012-01-01', '2020-01-01')}

A vector indicating the lower and upper bounds of the time series for longitudinal analyses}

\item{site_list}{\emph{list} || \strong{required}

A list of sites for which you would like to examine clinical facts. Can be one site
(single-site) or multiple (multi-site). Ensure that all sites listed here exist
in the provided cohort table.}
}
\value{
This function will return a dataframe where the output of the provided \code{check_func}
has been executed once for each \code{time_period} in the provided \code{time_span} for each of
the sites included in \code{site_list}
}
\description{
This function will loop through a time series and execute a user-provided
function against only the data within the time period of interest. It will
also ensure that only patients whose start date has passed but whose end date
has not passed are included in the analysis for any given time period. This means
that, for example, a patient who has a cohort period 1/1/2020 to 1/1/2025 would
be included when looking at the year of 2024, but they would not be included when
looking at the year of 2019 as they had not yet entered the cohort.
}
\examples{
\dontrun{
# for a function like...
 mock_function <- function(cohort,
                           input_tbl){

     test <- input_tbl \%>\%
        inner_join(cohort) \%>\%
        group_by(site) \%>\%
        summarise(n_row = n())

     return(test)
 }

# allow this function to be executed for each time period in a series
cohort <- dplyr::tibble('person_id' = c(1,2,3,4),
                        'site' = c('Site A', 'Site A', 'Site B', 'Site B'),
                        'start_date' = c('2012-01-01', '2014-01-10',
                                         '2015-03-05, '2020-01-07'),
                        'end_date' = c('2015-01-01', '2019-01-10',
                                       '2010-03-05, '2024-01-07'))

compute_fot(cohort = cohort,
            check_func = function(dat){
              mock_function(cohort = dat,
                            input_tbl = cdm_tbl('condition_occurrence'))
            },
            site_col = 'site',
            reduce_id = NULL,
            time_period = 'year',
            time_span = c('2012-01-01','2022-12-31'),
            site_list = c('Site A', 'Site B'))


}

}
