% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{compute_jaccard}
\alias{compute_jaccard}
\title{Jaccard Index Anomaly Detection}
\usage{
compute_jaccard(jaccard_input_tbl, var_col, omop_or_pcornet)
}
\arguments{
\item{jaccard_input_tbl}{\emph{tabular input} || \strong{required}

A table that contains at least \code{person_id}/\code{patid} and a variable column, and
where each row represents a unique instance where the variable occurred for
a given patient

Alternatively, it can be a list of all unique \code{person_id}/\code{patid} and
variable combinations}

\item{var_col}{\emph{string} || \strong{required}

The name of the column within \code{jaccard_input_table} that contains all
the variables that should be compared to each other in the similarity index}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}
}
\value{
A table with pairs of variables, labeled \code{concept1} and \code{concept2},
counts & proportions of patients with each concept individually
(\code{concept1_ct}, \code{concept1_prop}, \code{concept2_ct}, \code{concept2_prop}),
count of patients with BOTH concepts (\code{cocount}),
count of patients with EITHER concept (\code{concept_count_union}),
and the \code{jaccard_index} statistic
}
\description{
This function will compute the Jaccard Similarity Index for each combination of
two variables that occur within a specific patient's record. This function is compatible
with both the OMOP and PCORnet CDMs based on the user's selection.
}
\examples{
# create sample input with person identifier and variable associations
sample_input <- dplyr::tibble('person_id' = c(1,1,2,2,3,4,5,5),
                              'variable' = c('dx', 'drug', 'drug', 'dx',
                                             'drug', 'dx', 'dx', 'drug'))
# compute jaccard index
compute_jaccard(jaccard_input_tbl = sample_input,
                var_col = 'variable',
                omop_or_pcornet = 'omop')


}
