% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection_pcornet.R
\name{compute_jaccard_pcnt}
\alias{compute_jaccard_pcnt}
\title{Compute Jaccard Index}
\usage{
compute_jaccard_pcnt(jaccard_input_tbl, var_col)
}
\arguments{
\item{jaccard_input_tbl}{tbl that will undergo jaccard index computation;
the requirement is that it contains at least two columns: \code{person_id} and \code{concept_id}
where each row represents an instance where a specific \code{concept_id} is used for a given patient (\code{person_id})
Alternatively, it can be a list of all unique \code{person_id} and \code{concept_id} combinations}

\item{var_col}{the column within \code{jaccard_input_table} that contains all the concepts that should be compared to each other}
}
\value{
a table with both concepts, labeled \code{concept1} and \code{concept2}, the co-occurrence (\code{cocount}), individual
concept counts (\code{concept1_ct}, \code{concept2_ct}), total unique patient counts where either code is used (\code{concept_count_union}),
the \code{jaccard_index}, as well as proportion of patients where the concept appears (\code{concept1_prop}, \code{concepet2_prop})
}
\description{
Compute Jaccard Index
}
