% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Hotspots Anomaly Detection}
\usage{
detect_outliers(
  df_tbl,
  tail_input = "both",
  p_input = 0.9,
  column_analysis,
  column_variable,
  column_eligible = "analysis_eligible"
)
}
\arguments{
\item{df_tbl}{\emph{tabular input} || \strong{required}

A table meeting the previously described criteria. This input will
typically be the table output by \code{\link[=compute_dist_anomalies]{compute_dist_anomalies()}}}

\item{tail_input}{\emph{string} || defaults to \code{both}

A string indicating whether outliers are identified for positive values,
negative values, or both.

Acceptable inputs are \code{positive}, \code{negative}, or \code{both}}

\item{p_input}{\emph{numeric} || defaults to \code{0.9}

The p-value threshold that should be used to identify anomalies}

\item{column_analysis}{\emph{string} || \strong{required}

The name of the numerical column that is the target of the anomaly detection analysis}

\item{column_variable}{\emph{string} || \strong{required}

The name of the column with the variable to which the analysis column is related
(ex: concept_id for the Concept Set Distribution module)}

\item{column_eligible}{\emph{string} || defaults to \code{analysis_eligible}

The name of the column that indicates eligibility for analysis}
}
\description{
This function will identify anomalies in a dataframe using the
\code{\link[hotspots:hotspots]{hotspots::outliers()}} function. It assumes:
(1) No time component;
(2) Table has a column indicating whether a particular group or row is eligible for analysis;
(3) numerical variable exists that should be used for anomaly detection analysis
These conditions are met by the output of \code{\link[=compute_dist_anomalies]{compute_dist_anomalies()}}, which is typically
the input for this function
}
\examples{
# create sample input (modeled after EVP)
sample_ms_input <- dplyr::tibble('site' = c('Site A', 'Site A', 'Site A',
                                            'Site A', 'Site B', 'Site B',
                                            'Site B', 'Site B'),
                                 'variable' = c('dx', 'dx', 'drug', 'drug',
                                                'dx', 'dx', 'drug', 'drug'),
                                 'count' = c(100, 140, 39, 42, 137, 111,
                                             12, 15),
                                 'total_var' = c(1000, 1000, 200, 200, 1500,
                                                 1500, 100, 100))
# execute the full analysis, including compute_dist_anomalies and
# detect_outliers
anomaly_output1 <- compute_dist_anomalies(df_tbl = sample_ms_input,
                                          grp_vars = 'variable',
                                          var_col = 'count',
                                          denom_cols = 'total_var')

anomaly_output1

anomaly_output2 <- detect_outliers(df_tbl = anomaly_output1,
                                   column_analysis = 'count',
                                   column_variable = 'variable')

anomaly_output2

}
