% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{extract_color}
\alias{extract_color}
\title{Function to extract colors as hex codes}
\usage{
extract_color(...)
}
\arguments{
\item{...}{Character names of squba_colors_standard}
}
\value{
name and hex code/s for specified color
example useage: extract_color() (returns all colors) or extract_color("rust") (just returns "rust")
}
\description{
Function to extract colors as hex codes
}
