% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_summary.R
\name{make_interactive_squba}
\alias{make_interactive_squba}
\title{Make \code{squba} outputs interactive}
\usage{
make_interactive_squba(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{\emph{ggplot} || \strong{required}

A ggplot object output by any of the \verb{*_output} functions native
to each module. This will only work with graphs generated by \code{squba},
as it adds some additional metadata to each plot that is required for
this function to work.}
}
\value{
The same graph with interactive functionality and pre-set tooltips
based either in the \code{ggiraph} package or the \code{plotly} package
}
\description{
This function converts a ggplot object output by any of the *_output functions
into an interactive ggiraph or plotly object.
}
\examples{
\dontrun{
# first, execute the *_output function of interest to build a graph
grph_output <- *_output(process_output = my_df,
                        ...)

# then, run make_interactive_squba to activate interactive functionality
# some graphs may not have interactive abilities, which will be communicated
# via a message in the console

make_interactive_squba(grph_output)

}


}
