% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_summary.R
\name{make_interactive_ssdqa}
\alias{make_interactive_ssdqa}
\title{Make ggplot2 graphical output interactive}
\usage{
make_interactive_ssdqa(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot object output by any of the \verb{*_output} functions native
to each module}
}
\value{
the same graph with interactive functionality and pre-set tooltips
based either in the \code{ggiraph} package or the \code{plotly} package
}
\description{
Make ggplot2 graphical output interactive
}
