% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_summary.R
\name{param_csv_summ2}
\alias{param_csv_summ2}
\title{Generate parameter summary and recommended string to input into output function}
\usage{
param_csv_summ2(check_string, ...)
}
\arguments{
\item{check_string}{abbreviation to represent check type, should be the same as what
is prefixed to the names of the output functions}

\item{...}{all of the parameters input into the core function. any argument that is not
able to be vectorized (i.e. a CDM tbl, codeset, etc) will not appear in the final
summary}
}
\value{
paramater_summary.csv to the results directory

output_type string to be piped into a descriptive message at the end of the core function
to inform users what should be used as the \code{output_function} argument in the output
function
}
\description{
Generate parameter summary and recommended string to input into output function
}
