% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_prep.R
\name{prepare_cohort}
\alias{prepare_cohort}
\title{Prepare Cohort}
\usage{
prepare_cohort(cohort_tbl, age_groups = NULL, omop_or_pcornet)
}
\arguments{
\item{cohort_tbl}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer / character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}}

\item{age_groups}{\emph{tabular input} || defaults to \code{NULL}

If you would like to stratify the results by age group, create a table or
CSV file with the following columns and use it as input to this parameter:
\itemize{
\item \code{min_age} | \emph{integer} | the minimum age for the group (i.e. 10)
\item \code{max_age} | \emph{integer} | the maximum age for the group (i.e. 20)
\item \code{group} | \emph{character} | a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave as \code{NULL}}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data}
}
\value{
This function will return the same input cohort table with the addition of
the column \code{fu}, which includes the length of patient follow-up based on the
provided start & end dates. Optionally, if age_groups is not NULL, the output
will also include \code{age_ce} (patient age at cohort entry) and \code{age_grp}
(the user-provided age group category)
}
\description{
For the cohort table input into any of the *_process functions, this function
will prepare that cohort for the rest of the analysis. This includes computing
follow-up based on start and end dates and computing age at cohort entry categorized
into user-provided groups.
}
\examples{
\dontrun{
sample_cohort <- dplyr::tibble(site = c('Site A', 'Site B'),
                               person_id = c(1, 2),
                               start_date = c(as.Date('2012-01-01'),
                                              as.Date('2014-02-01)),
                               end_date = c(as.Date('2015-01-01'),
                                            as.Date('2020-02-01)))

age_grp_tbl <- dplyr::tibble(min_age = c(0, 7),
                             max_age = c(6, 10),
                             group = c('0-6', '7-10'))

prepare_cohort(cohort_tbl = sample_cohort,
               age_groups = age_grp_tbl,
               omop_or_pcornet = 'omop')

}


}
