% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_prep_pcornet.R
\name{prepare_cohort_pcnt}
\alias{prepare_cohort_pcnt}
\title{Prepare cohort for check execution
requirement: fields must have columns:
\code{person_id}, \code{start_date}, \code{end_date}}
\usage{
prepare_cohort_pcnt(cohort_tbl, age_groups = NULL, codeset = NULL)
}
\arguments{
\item{cohort_tbl}{table with required fields for each member of the cohort}

\item{age_groups}{option to read in a CSV with age group designations to allow for stratification
by age group in output. defaults to \code{NULL}.
sample CSV can be found in \code{specs/age_group_definitions.csv}}

\item{codeset}{option to read in a CSV with codeset metadata to allow for labelling of
cohort members based on a user-provided codeset. the codeset itself should be
a CSV file with at least a \code{concept_id} column and a \code{flag} column with user-provided
labels.
a sample metadata CSV, where the user can provide the correct table and column information,
can be found in \code{specs/codeset_metadata.csv}}
}
\value{
a tbl with person_id and the following:
\code{start_date} the cohort entry date
\code{end_date} the last visit
\code{fu}: length of follow up
\code{site} : patient site
if age_groups is not NULL:
\code{age_ce}: patient age at cohort entry
\code{age_grp}: user-provided age grouping
if codeset is not NULL:
\code{flag}: flag that indiciates patient is a member of a user-specified group in the codeset
}
\description{
Prepare cohort for check execution
requirement: fields must have columns:
\code{person_id}, \code{start_date}, \code{end_date}
}
