% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_site.R
\name{replace_site_col}
\alias{replace_site_col}
\title{Replace Summary Site Column}
\usage{
replace_site_col(tbl)
}
\arguments{
\item{tbl}{the tbl with to use for
replacement of \code{site_summ}}
}
\value{
tbl with \code{site} replacing \code{site_summ}
}
\description{
For analyses where a summary site column was created by \code{\link[=check_site_type]{check_site_type()}},
this function will replace the name of that column with the original "site"
name.
}
