% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_site_pcornet.R
\name{replace_site_col_pcnt}
\alias{replace_site_col_pcnt}
\title{check for \code{site_summ} column and switch to \code{site}}
\usage{
replace_site_col_pcnt(tbl)
}
\arguments{
\item{tbl}{the tbl with to use for
replacement of \code{site_summ}}
}
\value{
tbl with \code{site} replacing \code{site_summ}
}
\description{
check for \code{site_summ} column and switch to \code{site}
}
