% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_domain_file}
\alias{sample_domain_file}
\title{Sample Domain Definition File}
\format{
\subsection{\code{sample_domain_file}}{

A data frame with 3 rows and 7 columns
\describe{
\item{domain}{An arbitrary string label to refer to the domain}
\item{domain_tbl}{The name of the CDM table associated with the domain}
\item{source_col}{The column in the domain_tbl where source concepts are stored, if one exists. This is currently only used in the Souce and Concept Vocabularies Module.}
\item{concept_col}{The column in the domain_tbl where the primary concept of interest is stored}
\item{date_col}{The date column in the domain_tbl that should be used as the default for any date filtering operations}
\item{filter_logic}{(optional) A string indicating how the domain_tbl should be filtered to achieve the desired output. Should be written as if a \code{filter} command is being executed.}
\item{vocabulary_col}{(PCORnet only) The column in the domain_tbl where the type of vocabulary is indicated. If this column is used, please ensure concept sets have an associated \code{vocabulary_id} column to be used in the join.}
}
}
}
\usage{
sample_domain_file
}
\description{
A sample version of the file structure expected for domain
definition files that are used across several different modules.
The user should recreate this file structure and include their
domains of interest.
}
\details{
This is where the user defines which clinical domains are relevant
to their analysis and which columns should be used to conduct the
analysis. One file can be used across multiple modules and filtered down
to specific domains if needed.
}
\keyword{datasets}
