% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_squba}
\alias{scale_color_squba}
\title{squba Color Scale Constructor}
\usage{
scale_color_squba(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{\emph{string} || defaults to \code{main}

The name of the palette as it appears in \code{squba_palettes_standard}}

\item{discrete}{\emph{boolean} || defaults to \code{TRUE}

A boolean indicating whether the color aesthetic is discrete or not}

\item{reverse}{\emph{boolean} || defaults to \code{FALSE}

A boolean indicating whether the palette should be reversed}

\item{...}{Any additional arguments passed to \link[ggplot2:discrete_scale]{ggplot2::discrete_scale} or
\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn}, used respectively when
discrete is TRUE or FALSE}
}
\description{
This function will operate with ggplot objects to use the \code{squba}
standard colors/color palettes to add color to graphs
}
\examples{
library(ggplot2)

mtcars \%>\%
  tibble::rownames_to_column('car_model') \%>\%
  ggplot(aes(x = mpg, y = wt, color = car_model)) +
  geom_point() +
  scale_color_squba()

}
