% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_ssdqa}
\alias{scale_color_ssdqa}
\title{SSDQA Color Scale Constructor}
\usage{
scale_color_ssdqa(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ssdqa_palettes_standard.
If no palette specified, defaults to "main" palette}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
This function will operate with ggplot objects to use the SSDQA
standard colors/color palettes to add color to graphs
}
\examples{
library(ggplot2)

mtcars \%>\%
  tibble::rownames_to_column('car_model') \%>\%
  ggplot(aes(x = mpg, y = wt, color = car_model)) +
  geom_point() +
  scale_color_ssdqa()

}
