% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_squba}
\alias{scale_fill_squba}
\title{squba Fill Scale Constructor}
\usage{
scale_fill_squba(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{\emph{string} || defaults to \code{main}

The name of the palette as it appears in \code{squba_palettes_standard}}

\item{discrete}{\emph{boolean} || defaults to \code{TRUE}

A boolean indicating whether the fill aesthetic is discrete or not}

\item{reverse}{\emph{boolean} || defaults to \code{FALSE}

A boolean indicating whether the palette should be reversed}

\item{...}{Any additional arguments passed to \link[ggplot2:discrete_scale]{ggplot2::discrete_scale} or
\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn}, used respectively when
discrete is TRUE or FALSE}
}
\description{
This function will operate with ggplot objects to use the \code{squba}
standard colors/color palettes to add fill to graphs
}
\examples{
library(ggplot2)

mtcars \%>\%
  tibble::rownames_to_column('car_model') \%>\%
  ggplot(aes(x = car_model, y = wt, fill = car_model)) +
  geom_col() +
  scale_fill_squba()

}
