% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{squba_colors_standard}
\alias{squba_colors_standard}
\title{SQUBA Standard Color Hexes}
\format{
An object of class \code{character} of length 13.
}
\usage{
squba_colors_standard
}
\description{
SQUBA Standard Color Hexes
}
\keyword{datasets}
