% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{squba_palettes_standard}
\alias{squba_palettes_standard}
\title{SQUBA Standard Color Palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
squba_palettes_standard
}
\description{
SQUBA Standard Color Palettes
}
\keyword{datasets}
