% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ssdqa_colors_standard}
\alias{ssdqa_colors_standard}
\title{Setting standard colors}
\format{
An object of class \code{character} of length 12.
}
\usage{
ssdqa_colors_standard
}
\description{
Setting standard colors
}
\keyword{datasets}
