% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ssdqa_pal}
\alias{ssdqa_pal}
\title{Return function to interpolate a color palette}
\usage{
ssdqa_pal(palette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ssdqa_palettes_standard}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
color palettes, interpolated if necessary, with specified scheme and number of colors
example usage: ssdqa_pal("beachy")(10)
}
\description{
Return function to interpolate a color palette
}
