% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{anomalize_ss_anom_at}
\alias{anomalize_ss_anom_at}
\title{Run Anomilization for *_ss_anom_at}
\usage{
anomalize_ss_anom_at(fot_input_tbl, grp_vars, time_var, var_col)
}
\arguments{
\item{fot_input_tbl}{table output by compute_fot where the check of interest
is used as the check_func}

\item{grp_vars}{the variables that should be preserved in the cross join}

\item{time_var}{the column with time information}

\item{var_col}{the column with the numerical statistic of interest for the euclidean
distance computation}
}
\value{
one dataframe with all columns from the original input table
plus the columns needed for timetk output generated by the
\code{anomalize} function
}
\description{
Run Anomilization for *_ss_anom_at
}
