% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_prep.R
\name{build_birth_date}
\alias{build_birth_date}
\title{Build birth_date Column}
\usage{
build_birth_date(cohort, person_tbl)
}
\arguments{
\item{cohort}{\emph{tabular input} || \strong{required}

A table with cohort members, which contains at least a \code{person_id} column}

\item{person_tbl}{\emph{tabular input} || \strong{required}

The CDM \code{person} table or another table that contains at least \code{person_id},
\code{year_of_birth}, \code{month_of_birth}, and \code{day_of_birth}}
}
\value{
This function will return a table with a new birth_date column that is a
combination of year, month, and day of birth. This is intended to facilitate
computations using date of birth, like computing age.
}
\description{
This function, which is specific to the OMOP CDM implementation, will create
an aggregate birth_date column using \code{year_of_birth}, \code{month_of_birth}, and
\code{day_of_birth}
}
\examples{
\dontrun{
sample_cohort <- dplyr::tibble(site = c('Site A', 'Site B'),
                               person_id = c(1, 2),
                               start_date = c(as.Date('2012-01-01'),
                                              as.Date('2014-02-01)),
                               end_date = c(as.Date('2015-01-01'),
                                            as.Date('2020-02-01)))

build_birth_date(cohort = copy_to_new(df = sample_cohort),
                 person_tbl = cdm_tbl('person'))

}

}
