% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_prep.R
\name{calc_days_between_dates}
\alias{calc_days_between_dates}
\title{Calculate Date Differences in Multiple SQL Backends}
\usage{
calc_days_between_dates(
  date_col_1,
  date_col_2,
  db = get_argos_default()$config("db_src")
)
}
\arguments{
\item{date_col_1}{Date col 1}

\item{date_col_2}{Date col 2}

\item{db}{connection type object. Defaulted to config('db_src') for standard framework
Functionality added for Postgres, MS SQL and Snowflake}
}
\value{
an integer representing the difference (in days) between the two provided
dates
}
\description{
Function to get sql code for number of days between date1 and date2.
Adapted for sql dialects for Postgres, MSSQL, Snowflake, Oracle, BigQuery,
Presto, & Spark.
}
\details{
Should always be wrapped by sql()
}
\keyword{internal}
