% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_site.R
\name{check_site_type}
\alias{check_site_type}
\title{Check site type (single vs multi) against number of sites provided in list}
\usage{
check_site_type(cohort, multi_or_single_site)
}
\arguments{
\item{cohort}{a cohort tbl with a column for \code{site}, \code{person_id}, \code{start_date} and \code{end_date}}

\item{multi_or_single_site}{Option to run the function on a single vs multiple sites
- @single - run the function for a single site
- @multi - run the function for multiple sites}
}
\value{
if @multi_or_single_site = single but multiple sites are provided, the cohort table
is returned with a summary site column equaling \code{combined} so all sites will be treated
as one

\if{html}{\out{<div class="sourceCode">}}\preformatted{    otherwise, the existing site column is left alone. if an illogical parameter combination
    is supplied, the function will provide an error with recommendations on how to remedy the
    issue.
}\if{html}{\out{</div>}}
}
\description{
Check site type (single vs multi) against number of sites provided in list
}
