% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{compute_dist_anomalies}
\alias{compute_dist_anomalies}
\title{Check anomaly detection eligibility}
\usage{
compute_dist_anomalies(df_tbl, grp_vars, var_col, denom_cols)
}
\arguments{
\item{df_tbl}{output from the computation of a particular function for anomaly detection}

\item{grp_vars}{the columns to group by to compute the summary statistics for}

\item{var_col}{column to perform summary statistics for, to detect an anomaly}

\item{denom_cols}{denominator (or other) columns that should be preserved without nulls after
the cross_join takes place}
}
\value{
the \code{df_tbl} with the following computed:
\code{mean_val}, \code{median_val}, \code{sd_val}, \code{mad_val}, \code{cov_val}, \code{max_val},
\code{min_val}, \code{range_val}, \code{total_ct}, \code{analysis_eligible}
the \code{analysis_eligible} will indicate whether the group for which the user
wishes to detect an anomaly for is eligible for analysis.

The following conditions will disqualify a group from the anomaly detection analysis:
(1) Sample size < 5 in group
(2) Mean < 0.02 or Median < 0.01
(3) Mean value < 0.05 and range <0.01
(4) Coefficient of variance <0.01 and sample size <11
}
\description{
Check anomaly detection eligibility
}
