% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_detection.R
\name{compute_dist_mean_median}
\alias{compute_dist_mean_median}
\title{Compute Distance from Mean & Median}
\usage{
compute_dist_mean_median(tbl, grp_vars, var_col, num_sd, num_mad)
}
\arguments{
\item{tbl}{\emph{tabular input} || \strong{required}

A table with the target numerical variable (\code{var_col}) for the analysis
and each of the required grouping variables (\code{grp_vars})}

\item{grp_vars}{\emph{string or vector} || \strong{required}

The name of the variable(s) to group by when computing summary statistics}

\item{var_col}{\emph{string} || \strong{required}

The name of the numeric variable that should be the target of summary
statistic computations}

\item{num_sd}{\emph{integer} || \strong{required}

The number of standard deviations away from the mean that should be used
as the lower and upper bounds for outlier detection. Values falling, for
example, over 2 standard deviations above or below the mean would be
considered outliers.}

\item{num_mad}{\emph{integer} || \strong{required}

The number of median absolute deviations (MADs) away from the median that
should be used as lower and upper bounds. Outliers are not formally identified
based on the median, but the information will be available in the final
table should you prefer that method.}
}
\value{
A table where, for each group in \code{grp_vars}, various summary statistics like the
mean, median, standard deviation, MAD, and others, are computed based on the
\code{var_col}. Outliers are identified in the \code{anomaly_yn} column based on whether
the data point is +/- \code{num_sd} from the mean or the data point is > the 90th
percentile.
}
\description{
This function will, for the values in a given \code{var_col}, compute the
distance between that value and the mean & median values for each group
established by the \code{grp_vars} parameter.
}
\examples{
# sample input table
sample_input <- dplyr::tibble('variable' = c('scd', 'scd', 'scd', 'scd'),
                              'site' = c('Site A', 'Site A', 'Site B',
                                         'Site B'),
                              'count' = c(15, 24, 100, 93))

# execute function against sample data
compute_dist_mean_median(tbl = sample_input,
                         grp_vars = 'variable',
                         var_col = 'count',
                         num_sd = 1,
                         num_mad = 1)

}
