% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_summary.R
\name{generate_ref_table}
\alias{generate_ref_table}
\title{Create Concept Reference Table}
\usage{
generate_ref_table(tbl, id_col, name_col, denom, time = FALSE)
}
\arguments{
\item{tbl}{intermediate table generated in the output function that contains the concepts
of interest to be displayed in the reference table}

\item{id_col}{the name of the column with the concept that needs to be summarised in the
reference table}

\item{name_col}{the name of the column with the concept name associated with the concept in
\code{id_col}}

\item{denom}{the denominator count associated with @id_col to be displayed in the
reference table}

\item{time}{logical to define whether @tbl has over time output or not}
}
\value{
a reference table with summary information about the codes in the output that
could not be displayed in the associated graph
}
\description{
For several SQUBA modules, this function is used to create a summary reference
table that displays a more detailed breakdown of concept usage
}
\examples{
# generate reference table for non-time dependent concept summary

input_tbl_notime <- dplyr::tibble('concept_id' = c(1, 2, 3, 4),
                                  'concept_name' = c('test1', 'test2',
                                                     'test3', 'test4'),
                                  'ct_concept' = c(100, 200, 300, 400),
                                  'site' = c('Site A', 'Site A', 'Site A',
                                  'Site A'))

generate_ref_table(tbl = input_tbl_notime,
                   id_col = 'concept_id',
                   name_col = 'concept_name',
                   denom = 'ct_concept',
                   time = FALSE)

# generate reference table for time dependent concept summary

input_tbl_time <- dplyr::tibble('concept_id' = c(1, 2, 3, 4, 1, 2, 3, 4),
                                'time_start' = c('2012-01-01', '2012-01-01',
                                                 '2012-01-01', '2012-01-01',
                                                 '2013-01-01', '2013-01-01',
                                                 '2013-01-01', '2013-01-01'),
                                'time_increment' = c('year','year','year',
                                                     'year','year','year',
                                                     'year','year'),
                                'concept_name' = c('test1', 'test2', 'test3',
                                                   'test4', 'test1', 'test2',
                                                   'test3', 'test4'),
                                'ct_concept' = c(100, 200, 300, 400, 200,
                                                 300, 400, 500),
                                'site' = c('Site A', 'Site A', 'Site A',
                                           'Site A', 'Site A', 'Site A',
                                           'Site A', 'Site A'))

generate_ref_table(tbl = input_tbl_time,
                   id_col = 'concept_id',
                   name_col = 'concept_name',
                   denom = 'ct_concept',
                   time = TRUE)


}
