% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_session.R
\name{initialize_dq_session}
\alias{initialize_dq_session}
\title{Initialize Argos Session for DQ work}
\usage{
initialize_dq_session(
  session_name,
  db_conn,
  is_json = FALSE,
  cdm_schema,
  results_schema = NULL,
  results_tag = NULL
)
}
\arguments{
\item{session_name}{arbitrary string to name the session}

\item{db_conn}{either a connection object used to connect to a relational database OR the
path to a json file with the relevant connection information; if the latter,
is_json should be set to TRUE}

\item{is_json}{a logical to indicate whether db_conn is the path to a json file or not}

\item{cdm_schema}{string name of the schema where the data in a CDM format is kept}

\item{results_schema}{string name of the schema where results should be output if the user chooses
to utilize the results_tbl function native to the argos environment;
defaults to NULL}

\item{results_tag}{string to indicate a suffix (if any) that should be appended onto
any results tables; defaults to NULL}
}
\value{
Argos session will be established in the global environment; this session is
automatically established as the default

\if{html}{\out{<div class="sourceCode">}}\preformatted{    Function will print the database connection information and session
    information in the console for user review
}\if{html}{\out{</div>}}
}
\description{
Initialize Argos Session for DQ work
}
