% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_session.R
\name{initialize_dq_session}
\alias{initialize_dq_session}
\title{Initialize Argos Session for DQ}
\usage{
initialize_dq_session(
  session_name,
  db_conn,
  working_directory = getwd(),
  file_subdirectory,
  is_json = FALSE,
  cdm_schema,
  results_schema = NULL,
  vocabulary_schema = NULL,
  results_tag = NULL
)
}
\arguments{
\item{session_name}{arbitrary string to name the session}

\item{db_conn}{either a connection object used to connect to a relational database OR the
path to a json file with the relevant connection information; if the latter,
is_json should be set to TRUE}

\item{working_directory}{the base directory in which the analysis is taking place; defaults to the output of \code{\link[base:getwd]{base::getwd()}}}

\item{file_subdirectory}{the \emph{subdirectory} within the working directory where all files to be used in the analysis
(i.e. concept sets) are kept. this sets a default file location so the functions can easily
read in relevant files without having to redefine the path}

\item{is_json}{a logical to indicate whether db_conn is the path to a json file or not}

\item{cdm_schema}{string name of the schema where the data in a CDM format is kept}

\item{results_schema}{string name of the schema where results should be output if the user chooses
to utilize the \code{argos::results_tbl()} function native to the argos environment;
defaults to NULL}

\item{vocabulary_schema}{string name of the schema where vocabulary tables (i.e. concept) are
stored on the database}

\item{results_tag}{string to indicate a suffix (if any) that should be appended onto
any results tables; defaults to NULL}
}
\value{
Argos session will be established in the environment; this session is
automatically established as the default that will appear when \code{argos::get_argos_default()}
is called

\if{html}{\out{<div class="sourceCode">}}\preformatted{    Function will print the database connection information and session
    information in the console for user review
}\if{html}{\out{</div>}}
}
\description{
This is a wrapper function that will create an argos session and set up internal configurations
that will allow all downstream functions to access argos convenience functions.
This step is \emph{REQUIRED}, and makes it easier to connect to a backend database to conduct analyses.
The standard argos workflow can also be used, but this wrapper is provided for convenience.
}
\examples{
\dontrun{

## Create a database connection with DBI or input a file path
## to a json file with connection details
conn_dbi <- DBI::dbConnect(drv = my_driver_func(), # insert appropriate driver
                           dbname = "my_dbname",
                           host = "my_host",
                           port = "my_port",
                           user = "my_username",
                           password = "my_password")

conn_json <- "path/to/connection/file"

## Establish session and load appropriate convenience functions &
## configurations into the environment

initialize_dq_session(session_name = "my_session",
                      db_conn = conn_dbi,
                      is_json = FALSE,
                      working_directory = get_wd(),
                      file_subdirectory = "my_files",
                      cdm_schema = "my_schema")

}


}
