% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_summary.R
\name{join_to_vocabulary}
\alias{join_to_vocabulary}
\title{Join to OMOP Vocabulary Table}
\usage{
join_to_vocabulary(tbl, vocab_tbl, col, vocab_col = "concept_id")
}
\arguments{
\item{tbl}{data table to which the vocabulary table should be joined}

\item{vocab_tbl}{location of the vocabulary table}

\item{col}{the column that should be used in the \code{by} statement to join
to the vocab_col in the vocabulary table}

\item{vocab_col}{the column in the vocabulary table that should be used to join
to the \code{tbl}}
}
\value{
the dataframe provided in \code{tbl} with the addition of the concept name
column
}
\description{
This is a convenience function that allows users to join to an OMOP
vocabulary table (most likely the \code{concept} table)
}
\examples{
\dontrun{

sample_input_tbl <- dplyr::tibble('concept' = c(1234, 5678, 91011),
                                  'ct_concept' = c(100, 200, 300),
                                  'site' = c('Site A', 'Site A', 'Site A'))

join_to_vocabulary(tbl = sample_input_tbl,
                   vocab_tbl = vocabulary_tbl('concept'),
                   col = 'concept',
                   vocab_col = 'concept_id')
}


}
