% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_summary.R
\name{join_to_vocabulary}
\alias{join_to_vocabulary}
\title{Join to vocabulary table}
\usage{
join_to_vocabulary(tbl, vocab_tbl, col, vocab_col = "concept_id")
}
\arguments{
\item{tbl}{data table to which the vocabulary table should be joined}

\item{vocab_tbl}{location of the vocabulary table}

\item{col}{the column that should be used in the \code{by} statement to join
to the vocab_col in the vocabulary table}

\item{vocab_col}{the column in the vocabulary table that should be used to join
to the @tbl}
}
\value{
the dataframe provided in @tbl with the addition of the concept name
column
}
\description{
Join to vocabulary table
}
