% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_loops.R
\name{loop_through_visits}
\alias{loop_through_visits}
\title{Patient Facts per Visit Type}
\usage{
loop_through_visits(
  cohort_tbl,
  omop_or_pcornet,
  check_func,
  site_col,
  time = FALSE,
  visit_type_tbl,
  visit_tbl,
  site_list,
  visit_list = c("outpatient", "inpatient"),
  domain_tbl
)
}
\arguments{
\item{cohort_tbl}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}

Note that the start and end dates included in this table will be used to
limit the search window for the analyses in this module. It is recommended that
this table has been passed through the \link{prepare_cohort} function as well.}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data
\itemize{
\item \code{omop}: run the \code{\link[=loop_through_visits_omop]{loop_through_visits_omop()}} function against an OMOP CDM instance
\item \code{pcornet}: run the \code{\link[=loop_through_visits_pcnt]{loop_through_visits_pcnt()}} function against a PCORnet CDM instance
}}

\item{check_func}{\emph{function} || \strong{required}

The function that should be executed within each time period in the loop. This
parameter should be structured as the following, where \code{cht} is the cohort table
and \code{t} is the input data for the function:

\code{function(cht, t){check_function(param1 = cht, param2 = t, param3 = param3_input, ..., paramX = paramX_input)}}

Make sure to include all parameters required for the original function where
default values are not being used.}

\item{site_col}{\emph{string} || \strong{required}

The name of the column in the cohort table that contains the site names. This
will typically be either \code{site} or \code{site_summ}}

\item{time}{\emph{boolean} || defaults to \code{FALSE}

A boolean to indicate whether to execute a longitudinal analysis}

\item{visit_type_tbl}{\emph{tabular input} || \strong{required}

A table that defines visit types of interest called in \code{visit_list.} This input
should contain:
\itemize{
\item \code{visit_concept_id} / \code{visit_detail_concept_id} or \code{enc_type} | \emph{integer} or \emph{character} | the \verb{visit_(detail)_concept_id} or \code{enc_type} that represents the visit type of interest (i.e. 9201 or IP)
\item \code{visit_type} | \emph{character} | the string label to describe the visit type
}}

\item{visit_tbl}{\emph{tabular input} || defaults to \code{cdm_tbl('visit_occurrence')}

The CDM table with visit information (i.e. visit_occurrence or encounter)}

\item{site_list}{\emph{list} || \strong{required}

A list of sites for which you would like to examine clinical facts. Can be one site
(single-site) or multiple (multi-site). Ensure that all sites listed here exist
in the provided cohort table.}

\item{visit_list}{\emph{string or vector} || defaults to \code{c('outpatient', 'inpatient')}

A string or vector of visit types by which the output should be stratified. Each
visit type listed in this parameter should match an associated visit type defined
in the \code{visit_type_tbl}}

\item{domain_tbl}{\emph{tabular input} || \strong{required}

A table that defines the fact domains to be investigated in the analysis. This
input should contain:
\itemize{
\item \code{domain} | \emph{character} | a string label for the domain being examined (i.e. prescription drugs)
\item \code{domain_tbl} | \emph{character} | the CDM table where information for this domain can be found (i.e. drug_exposure)
\item \code{filter_logic} | \emph{character} | logic to be applied to the domain_tbl in order to achieve the definition of interest; should be written as if you were applying it in a dplyr::filter command in R
}}
}
\value{
This function will return a list of dataframes with the median number of facts per
patient for all domains in \code{domain_tbl}, where each dataframe is specific to a
given visit type from \code{visit_list}
}
\description{
This function will loop through each site and visit type provided and execute
the provided check function to compute facts stratified by visit type. Primarily intended
for use in the Patient Facts module, but can be adapted for use elsewhere.
}
\examples{
\dontrun{
# for a function like...
 mock_function <- function(cohort,
                           input_tbl){

     test <- input_tbl \%>\%
        inner_join(cohort) \%>\%
        group_by(site) \%>\%
        summarise(n_row = n())

     return(test)
 }

# allow this function to be executed while stratifying by visit type
cohort <- dplyr::tibble('person_id' = c(1,2,3,4),
                        'site' = c('Site A', 'Site A', 'Site B', 'Site B'),
                        'start_date' = c('2012-01-01', '2014-01-10',
                                         '2015-03-05, '2020-01-07'),
                        'end_date' = c('2015-01-01', '2019-01-10',
                                       '2010-03-05, '2024-01-07'))

sample_visits <- dplyr::tibble('visit_concept_id' = 9202,
                               'visit_type' = 'outpatient')

sample_domains <- dplyr::tibble('domain' = 'diagnosis',
                                'domain_tbl' = 'condition_occurrence',
                                'filter_logic' = NA)

loop_through_visits(cohort_tbl = cohort,
                    omop_or_pcornet = 'omop',
                    check_func = function(cht, t){
                       mock_function(cohort = cht,
                                     input_tbl = t)
                    },
                    site_col = 'site',
                    time = FALSE,
                    visit_type_tbl = sample_visits,
                    visit_tbl = cdm_tbl('visit_occurrence'),
                    site_list = c('Site A', 'Site B'),
                    visit_list = c('outpatient'),
                    domain_tbl = sample_domains)


}

}
