% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_loops.R
\name{loop_through_visits}
\alias{loop_through_visits}
\title{loops through visit types and sites to compute patient facts}
\usage{
loop_through_visits(
  cohort_tbl,
  check_func,
  site_col,
  time = FALSE,
  visit_type_tbl = read_codeset("pf_visit_types", "ic"),
  visit_tbl = cdm_tbl("visit_occurrence"),
  site_list,
  visit_list = c("inpatient", "outpatient"),
  grouped_list = c("person_id", "start_date", "end_date", "fu", "site"),
  domain_tbl = read_codeset("pf_domains_short", "cccc")
)
}
\arguments{
\item{cohort_tbl}{the tbl that comes from \code{prepare_pf}}

\item{check_func}{the base function for the check that needs to be executed across time; this argument
should be structured as the following, where cht is the cohort and t is the input data
for the function:

\if{html}{\out{<div class="sourceCode">}}\preformatted{              function(cht, t)\{check_function(param1 = cht, param2 = t, param3 = param3_input, ...,
              paramX = paramX_input)\}

              all parameters for the base check function should be included if any defaults are not being
              used
}\if{html}{\out{</div>}}}

\item{site_col}{the column in the data where the site variable can be found}

\item{time}{a logical indicating whether the analysis is being conducted longitudinally}

\item{visit_type_tbl}{The visit_concept_ids of interest for the analysis. \code{all} may be used in this field
to select every visit type; defaults to \code{pf_visit_types} in specs folder}

\item{visit_tbl}{the cdm visit_occurrence tbl; defaults to \code{cdm_tbl('visit_occurrence')}}

\item{site_list}{the sites to iterate through}

\item{visit_list}{the list of visit types to iterate through}

\item{grouped_list}{the input for which to group variables}

\item{domain_tbl}{defaults to \code{pf_domains} in the specs folder;
@domain: the domain name; output will have this domain
@default_tbl: the table to pull from
@field_name the field name to filter by; leave null if no filter
@field_filter: the filtered codes}
}
\value{
a returned list stratified by visit type
}
\description{
loops through visit types and sites to compute patient facts
}
