% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_summary.R
\name{make_interactive_ssdqa}
\alias{make_interactive_ssdqa}
\title{Create Interactive Graphical Output}
\usage{
make_interactive_ssdqa(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot object output by any of the \verb{*_output} functions native
to each module}
}
\value{
the same graph with interactive functionality and pre-set tooltips
based either in the \code{ggiraph} package or the \code{plotly} package
}
\description{
This function converts a ggplot object output by any of the *_output functions
into an interactive ggiraph or plotly object.
}
