% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_prep.R
\name{prepare_cohort}
\alias{prepare_cohort}
\title{Prepare Cohort}
\usage{
prepare_cohort(cohort_tbl, age_groups = NULL, codeset = NULL, omop_or_pcornet)
}
\arguments{
\item{cohort_tbl}{table with required fields for each member of the cohort}

\item{age_groups}{option to read in a table with age group designations to allow for stratification
by age group in output. defaults to \code{NULL}.}

\item{codeset}{option to read in a table with codeset metadata to allow for labelling of
cohort members based on a user-provided codeset. the codeset itself should be
a table with at least a \code{concept_id} column and a \code{flag} column with user-provided
labels.}

\item{omop_or_pcornet}{string indicating the appropriate backend CDM (\code{omop} or \code{pcornet})}
}
\value{
a tbl with person_id and the following:
\code{start_date} the cohort entry date
\code{end_date} the last visit
\code{fu}: length of follow up
\code{site} : patient site
if age_groups is not NULL:
\code{age_ce}: patient age at cohort entry
\code{age_grp}: user-provided age grouping
if codeset is not NULL:
\code{flag}: flag that indiciates patient is a member of a user-specified group in the codeset
}
\description{
For the cohort table input into any of the *_process functions, this function
will prepare that cohort for the rest of the analysis. This includes computing
follow-up based on start and end dates and computing age at cohort entry categorized
into user-provided groups. Note that the cohort table must have columns:
\code{site}, \code{person_id}, \code{start_date}, \code{end_date}
}
\examples{
\dontrun{
sample_cohort <- dplyr::tibble(site = c('Site A', 'Site B'),
                               person_id = c(1, 2),
                               start_date = c(as.Date('2012-01-01'),
                                              as.Date('2014-02-01)),
                               end_date = c(as.Date('2015-01-01'),
                                            as.Date('2020-02-01)))

age_grp_tbl <- dplyr::tibble(min_age = c(0, 7),
                             max_age = c(6, 10),
                             group = c('0-6', '7-10'))

prepare_cohort(cohort_tbl = sample_cohort,
               age_groups = age_grp_tbl,
               omop_or_pcornet = 'omop')

}


}
