% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_site.R
\name{replace_site_col}
\alias{replace_site_col}
\title{Replace Summary Site Column}
\usage{
replace_site_col(tbl)
}
\arguments{
\item{tbl}{\emph{tabular input} || \strong{required}

The table with a \code{site_summ} column that needs to be replaced}
}
\value{
The same table that was used as input, with \code{site} replacing \code{site_summ}
}
\description{
For analyses where a summary site column was created by \code{\link[=check_site_type]{check_site_type()}},
this function will replace the name of that column with the original "site"
name.
}
\examples{
## Sample input table
input_sample <- dplyr::tibble(person_id = c(1,2,3),
                              site_summ = c("combined","combined","combined"))

## Replace site_summ col for final output
replace_site_col(tbl = input_sample)

}
