% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_squba}
\alias{scale_fill_squba}
\title{SQUBA Fill Scale Constructor}
\usage{
scale_fill_squba(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in squba_palettes_standard
If no palette specified, defaults to "main" palette}

\item{discrete}{Boolean indicating whether fill aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
This function will operate with ggplot objects to use the SQUBA
standard colors/color palettes to add fill to graphs
}
\examples{
library(ggplot2)

mtcars \%>\%
  tibble::rownames_to_column('car_model') \%>\%
  ggplot(aes(x = car_model, y = wt, fill = car_model)) +
  geom_col() +
  scale_fill_squba()

}
