% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ssdqa_palettes_standard}
\alias{ssdqa_palettes_standard}
\title{SSDQA Standard Color Palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
ssdqa_palettes_standard
}
\description{
SSDQA Standard Color Palettes
}
\keyword{datasets}
